Working code
Got rid of the eval funcs class and just added that one line of code to the script loop.  Fixed problem where I was getting the same value for every run iteration of a path.  Fix was made by changing the expr assignment in the createFunction class on line 110 to np.copy().