Code working 4 analyses
Have the script working to get through all four analyses.  Need to track the progression of the path_vals matrix to make sure values calculated are expected based off initial values.  Need to consolidate this code for a third for loop so it is not repetitive.  Need to add another possible checker for a large rework loop where it is stated in the script.