Many Changes
Changed sequence for loop to a while loop so the sequence can be restarted.  Added an lrestart counter to go along with L1 and L4 loop counters.  Trying to collect all of the rework results I will want to plot.  And some other changes.  Will need to start working on normal RVs next for L4 rework and L restart